requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-socketio==5.11.4
httpx==0.27.2
slowapi==0.1.9
python-dotenv==1.0.1
email-validator==2.2.0
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
